:sparkles: Add method to create Reuse Notes PDF report
Add method reportReuseNotes in file ReportPDF, this method generate a PDF that displays a list of students who may have their grades reused based on the condition in the 'if' on line 82, futhermore, created two methods in ToDoModel to be used in 'if' on line 82, they are: 'averegeNotesValidator' and 'averageNote'. and lastly, added a method getDescriptionTg em file TGModel. The isApt in file DisplayTableModel was changed to public.